requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
openpyxl==3.0.9
requests==2.26.0
python-dotenv==0.19.1
werkzeug==2.0.1
gunicorn==20.1.0
sqlalchemy==1.4.23